Extract the data & do the brute-force time check
to save the data to a text file. Because sometimes we need the data as
a txt file for some programs that are NOT able to read the nc extension.

The line that saves the actual data is commented out in order to check
the performance using the `brute-force` method. This method uses the
difference between time-stamps before and after the process we need to
time. Then, the result is simply printed on the page.

I've done this in order to compare the performance using the `np.array`
methods instead of using `for` loops. The methods `repeat` and `tile`
were the ones I was looking for. I don't know if I'll try reaching the
same result using `for` loops or other stuff. I just assume that using
the native `numpy` functions have better performance than `for` loops.