Write a program to plot skin temperature
I'd like to make this completely automatic. Just give it an argument and
watch it plot the data as you please. Frame by frame, etc.

Later, this program has to plot the annual temperature difference for
each month available.

Also, the legends and the scales have to be identical for better
comparability.